Small Fixes & Features (#378)

+++ fastvideo/v1/tests/encoders/test_clip_encoder.py
@@ -66,7 +66,7 @@ def test_clip_encoder():
 
     # Load the HuggingFace implementation directly
     # model2 = CLIPTextModel(hf_config)
-    model2 = model2.to(torch.float16)
+    # model2 = model2.to(torch.float16)
     model2 = model2.to(device)
     model2.eval()
